Consolidate project synchronization
