+++ crates/djls-semantic/src/project.rs
@@ -1,22 +1,19 @@
-mod cache;
 mod db;
-mod django;
-mod external;
-mod indexes;
 #[path = "project/project.rs"]
 mod input;
 mod introspector;
 mod names;
 mod python;
-mod refresh;
 mod resolve;
 mod symbols;
+mod sync;
 
 pub use db::Db;
 pub use input::load_env_file;
+pub(crate) use input::project_model_modules;
+pub(crate) use input::project_templatetag_modules;
 pub use input::Project;
 pub use input::ProjectPythonIndex;
-pub(crate) use input::ProjectPythonModule;
 pub use input::ProjectTemplateFiles;
 pub use input::TemplateDirs;
 pub use introspector::ProjectIntrospector;
@@ -24,8 +21,6 @@ pub use names::LibraryName;
 pub use names::PyModuleName;
 pub use names::TemplateSymbolName;
 pub use python::Interpreter;
-pub use refresh::refresh_external_data;
-pub use symbols::load_template_library_cache;
 pub(crate) use symbols::DiscoveredSymbolCandidate;
 pub use symbols::InstalledSymbolCandidate;
 pub use symbols::InstalledSymbolOrigin;
@@ -38,3 +33,5 @@ pub use symbols::TemplateLibrarySnapshot;
 pub use symbols::TemplateSymbol;
 pub use symbols::TemplateSymbolKind;
 pub use symbols::TemplateSymbolSnapshot;
+pub use sync::load_template_library_cache;
+pub use sync::refresh_external_data;
